software/requirements.txt
numpy==1.21.6
opencv-contrib-python==4.5.2.52
paho-mqtt==1.6.1
picamera==1.13
Pillow==9.4.0
PyAudio==0.2.13
pyserial==3.5
RPi.GPIO==0.7.1
